requirements.txt
# Web Scraping Framework
scrapy==2.11.2
lxml==4.9.4

# Apache Kafka - Message Streaming
confluent-kafka==2.3.0

# PDF Text Extraction & OCR
pdfplumber==0.10.3
PyMuPDF==1.23.8
easyocr==1.7.0
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3

# Image Processing
Pillow==10.0.0

# Environment Variables Management
python-dotenv==1.0.0

# HTTP Library for API Requests
requests==2.31.0

#for database
psycopg2-binary==2.9.11

#for parsing with AI

groq